storage: ADR-034 widen InputIdx int32 → int64 — fixes 1B Reference bulk-load
The first 1B Reference + Sorted bulk-load run produced 213M atoms but an
empty GSPO index. Root cause: int32 overflow in the input-occurrence index.
At 1B triples × 4 atoms = 4B occurrences, the `int globalIdx` counter wrapped
past int32 max (~2.1B) to negative values around the half-billion-triple
mark. Negative InputIdx values broke the dense-coverage invariant in
DiskBackedAssignedIds.DiskBackedReader, which threw
InvalidOperationException ("dense-coverage violation") mid-way through
EnumerateResolved. The exception aborted FlushToDisk before _bulkSorter
drained into GSPO. The CLI exited without printing "Load complete" because
the unhandled exception killed the process post-vocabulary-write but
pre-summary.

Fix: widen int32 → int64 throughout the input-occurrence index path.

- SortedAtomBulkBuilder._globalIdx: int → long
- SortedAtomBulkBuilder._spillBuffer: List<(byte[], int)> → List<(byte[], long)>
- SortedAtomStoreExternalBuilder.SpillChunks: returns long, uses long globalIdx
- SortedAtomStoreExternalBuilder.SpillOneChunk: takes List<(byte[], long)>
- SortedAtomStoreExternalBuilder.MergeAndWrite: takes long inputCount
- SortedAtomStoreExternalBuilder.ChunkReader: returns long InputIdx
- Chunk file format header: int32 length + int32 InputIdx (8 bytes)
                          → int32 length + int64 InputIdx (12 bytes)
- In-memory assigned[] still uses int indexer with explicit checked cast.
  In-memory path is bounded by int32 array size anyway; explicit guard
  throws ArgumentOutOfRangeException with a useful message pointing at
  useDiskBackedAssigned.

Regression test (DiskBacked_ChunkFile_RoundTripsInt64InputIdx) verifies
the chunk-file format round-trips InputIdx values past int32 max. Going
through the bulk-builder at 4B+ occurrences would take hours; this isolates
the file-format byte-layout and arithmetic without that runtime.

Lesson: scale-boundary tests aren't optional. The unit suite at 20K triples
(80K occurrences) sits comfortably inside int32. The 100M production
gradient (400M occurrences) sits comfortably inside int32. Only at 1B
production (4B occurrences) does the boundary bite. We should have either
(a) tested explicitly past the boundary, (b) widened pre-emptively after
recognizing the architecture targets full Wikidata (85B occurrences).
Recorded as a session lesson for the gradient-discipline pattern.

511/511 Storage suite green.

+++ src/Mercury/Storage/SortedAtomBulkBuilder.cs
@@ -59,10 +59,10 @@ internal sealed class SortedAtomBulkBuilder : IDisposable
     // ADR-034 Phase 1B-5e streaming-input state. The buffer accumulates atom occurrences
     // until it crosses the chunk threshold, then sorts + spills to a chunk file. Memory
     // is bounded by chunkBufferBytes regardless of input scale.
-    private readonly List<(byte[] Bytes, int InputIdx)> _spillBuffer = new();
+    private readonly List<(byte[] Bytes, long InputIdx)> _spillBuffer = new();
     private long _spillBufferBytes;
     private readonly List<string> _chunkFiles = new();
-    private int _globalIdx;  // monotonic input-occurrence index across all atoms
+    private long _globalIdx;  // monotonic input-occurrence index across all atoms; widened from int to long for >500M-triple bulk loads (1B triples × 4 atoms = 4B occurrences exceeds int32)
     private long _tripleCount;
 
     // Disk-backed AssignedIds resolver (Phase 1B-5d). Allocated lazily in EnsureSpillerInitialized
